[FFRESW][project-files][libraries][ino]Updated FFRESW.ino file, fixed and adapted temperature as well as sensorModule to support new I2C MCP9601 boards and sensors. Used Flashstrings where possible.

+++ eSW/libraries/logManager/logManager.h
@@ -84,7 +84,7 @@ private:
     String baseLogFileName;
     
     static const int chipSelectPinEth = 10; // Default CS pin for SD card
-    static const int maxLogFileSize = 1024 * 1024 * 100; // 100 MB
+    static const long maxLogFileSize = 1024 * 1024 * 100; // 100 MB
 
     LogManager(const LogManager&) = delete;
     LogManager& operator=(const LogManager&) = delete;
